Pull out bootApplication into separate file
